backend/requirements.txt
Flask==2.3.0
Flask-Cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.1
bcrypt==4.1.2
PyJWT==2.8.0
python-socketio 
eventlet
whisper
pydub
flask-socketio